pyproject.toml

[project]
name = "mysingle"
version = "2.0.0-alpha"
description = "Unified utilities and gRPC protocols for MySingle Platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Pydantic (기본)
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    # Core 통합 모듈 의존성
    "structlog>=23.2.0",         # logging
    "colorlog>=6.9.0",           # logging
    "prometheus-client>=0.19.0", # metrics
    "motor>=3.3.2",              # base (Beanie)
    "beanie>=1.23.6",            # base (Beanie)
    "emails>=0.6",               # email
    "jinja2>=3.1.6",             # email
    "httpx>=0.25.2",             # audit
    "fastapi>=0.104.1",          # core.app_factory
    "uvicorn[standard]>=0.24.0", # core.app_factory
    "email-validator>=2.0.0",    # config (EmailStr validation)
    "pandas>=2.3.3",
]

[project.optional-dependencies]
# 인증
auth = ["pyjwt>=2.10.1", "pwdlib[argon2,bcrypt]>=0.2.1", "httpx-oauth>=0.16.1"]

# 웹 (FastAPI) - 이미 core에 포함되어 있으므로 빈 배열
web = []

# 데이터베이스 (추가 도구)
database = ["duckdb>=1.1.0", "redis>=6.4.0"]

# DSL
dsl = ["RestrictedPython>=7.0", "pandas>=2.2.0", "numpy>=1.26.0"]

# gRPC
grpc = ["grpcio>=1.60.0", "grpcio-tools>=1.60.0", "protobuf>=4.25.0"]

# 클라이언트
clients = ["aiohttp>=3.9.4"]

# 조합형 (자주 사용)
common = ["mysingle[auth,database,web]"]
common-grpc = ["mysingle[auth,database,web,grpc,clients]"]
full = ["mysingle[auth,web,database,dsl,grpc,clients]"]

# 개발 도구
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.26.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
    "pre-commit>=4.3.0",
]

[project.scripts]
mysingle-cli = "mysingle.cli.__main__:main"
mysingle-proto = "mysingle.cli.protos.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mysingle"]

[dependency-groups]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "mongomock>=4.1.2",
    "httpx>=0.26.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
    "pre-commit>=4.3.0",
    "bandit[toml]>=1.8.6",
    # gRPC and protobuf for testing
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    "googleapis-common-protos>=1.63.0",
    # Additional testing dependencies
    "pyjwt>=2.10.1",
    "pwdlib[argon2,bcrypt]>=0.2.1",
    "duckdb>=1.1.0",
    "RestrictedPython>=7.0",
    "numpy>=1.26.0",
    "mongomock-motor>=0.0.36",
]

# Ruff configuration (replaces Black + Flake8 + isort)
[tool.ruff]
line-length = 88
target-version = "py312"
exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "data/",
    "logs/",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by Black)
    "F403",   # 'from module import *' used; unable to detect undefined names (schemas 패턴)
    "B008",   # Do not perform function calls in argument defaults (FastAPI Depends/Query pattern)
    "B006",   # mutable-argument-default (리스트 기본값, 의도적 사용)
    "B904",   # raise ... from err (not always necessary for HTTP exceptions)
    "SIM105", # suppressible-exception (try-except-pass 패턴, 일부 허용)
    "SIM117",
    "SIM103",
]

[tool.ruff.lint.isort]
known-first-party = ["mysingle"]

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv", "__pycache__"]
skips = [
    "B101", # assert_used
    "B105", # hardcoded_password_string (상수 문자열)
    "B106", # hardcoded_password_funcarg (함수 인자 기본값)
    "B107", # hardcoded_password_default (기본값)
    "B601", # shell=True
    "B608", # hardcoded_sql_expressions (파라미터화된 쿼리 사용 중)
    # 데모/시뮬레이션 데이터 생성용 random 사용 (보안상 중요하지 않음)
    "B311", # Standard pseudo-random generators (dashboard_service.py, portfolio_service.py Mock 데이터)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "e2e: End-to-end integration tests (deselect with '-m \"not e2e\"')",
    "slow: Slow-running tests (deselect with '-m \"not slow\"')",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false           # 점진적 타입 도입
ignore_missing_imports = true
exclude = ["tests/", "build/", "dist/"]
# Temporarily disable no-any-return for gradual type improvement
disable_error_code = ["no-any-return"]
